python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "pay2play-arc"
version = "0.1.0"
description = "Python SDK for pay2play on Circle Arc — Vyper contracts + Circle Gateway x402"
requires-python = ">=3.11"
dependencies = [
  "vyper~=0.4.0",
  "titanoboa>=0.2.0",
  "eth-account>=0.11.0",
  "httpx>=0.27.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
circle = [
  "circle-titanoboa-sdk",
]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.24",
  "pytest-httpx>=0.32",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
